+++ Pod/Classes/Common/Logging.swift
@@ -8,11 +8,9 @@
 
 import Foundation
 
-private let marker = "[NSURLSession-Mock]"
-
 /**
  Wraps NSLog and makes sure we have a consistent marker in the console to filter by
  */
 func Log(format: String) {
-    NSLog("%@", "\(marker) \(format)")
+    NSLog("[NSURLSession-Mock] %@", format)
 }