+++ Pod/Classes/Common/MockResponse.swift
@@ -9,21 +9,16 @@
 import Foundation
 
 /**
- Describes a single response to a mocked request
-*/
-struct MockResponse {
-    let body: NSData?
-    let statusCode: Int
-    let headers: [String:String]
-
-    init(body: NSData?, statusCode: Int = 200, headers: [String:String] = [:]) {
-        self.body = body
-        self.statusCode = statusCode
-        self.headers = headers
-    }
+ Return one of this enum from a MockResponseHandler
+ */
+public enum MockResponse {
+    case Success(statusCode: Int, headers: [String:String], body: NSData?)
+    case Failure(error: NSError)
 }
 
 /**
  Given the URL and the extracted sections, what should be the response data, the status code and the headers.
+ 
+ Or, alternately, return a .Failure to deal with mocking errors before we hit the server (i.e. networking failure etc)
 */
-typealias MockResponseHandler = (url: NSURL, extractions: [String]) -> MockResponse
+public typealias MockResponseHandler = (url: NSURL, extractions: [String]) -> MockResponse