+++ Pod/Classes/NSURLSession/NSURLSession+Mock.swift
@@ -22,12 +22,6 @@ public enum RequestDebugLevel: Int {
     case All
 }
 
-/**
- Mocks which are interested in sections of the URL to create the response body
- should pass in functions matching this signature to `mockSingle` or `mockEvery`
-*/
-public typealias BodyFunction = [String] -> NSData
-
 extension NSURLSession {
     
     internal static let register = MockRegister()
@@ -43,9 +37,9 @@ extension NSURLSession {
      */
     public class func mockSingle(request: NSURLRequest, body: NSData?, headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay) {
         let matcher = SimpleRequestMatcher(url: request.URL!, method: request.HTTPMethod!)
-        self.mockSingle(matcher, response: { _ in MockResponse(body: body, statusCode: statusCode, headers: headers) }, delay: delay)
+        self.mockSingle(matcher, delay: delay) { _ in .Success(statusCode: statusCode, headers: headers, body: body) }
     }
-    
+
     /**
      All calls exactly matching `request` will successfully return `body`
      
@@ -57,7 +51,7 @@ extension NSURLSession {
      */
     public class func mockEvery(request: NSURLRequest, body: NSData?, headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay) {
         let matcher = SimpleRequestMatcher(url: request.URL!, method: request.HTTPMethod!)
-        self.mockEvery(matcher, response: { _ in return MockResponse(body: body, statusCode: statusCode, headers: headers) }, delay: delay)
+        self.mockEvery(matcher, delay:delay) { _ in return .Success(statusCode: statusCode, headers: headers, body: body) }
     }
     
     /**
@@ -72,7 +66,7 @@ extension NSURLSession {
      */
     public class func mockSingle(expression: String, HTTPMethod: String = "GET", body: NSData?, headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay) throws {
         let matcher = try SimpleRequestMatcher(expression: expression, method: HTTPMethod)
-        self.mockSingle(matcher, response: { _ in return MockResponse(body: body, statusCode: statusCode, headers: headers) }, delay: delay)
+        self.mockSingle(matcher, delay: delay) { _ in return .Success(statusCode: statusCode, headers: headers, body: body) }
     }
 
     /**
@@ -87,39 +81,39 @@ extension NSURLSession {
      */
     public class func mockEvery(expression: String, HTTPMethod: String = "GET", body: NSData?, headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay) throws {
         let matcher = try SimpleRequestMatcher(expression: expression, method: HTTPMethod)
-        self.mockEvery(matcher, response: { _ in return MockResponse(body: body, statusCode: statusCode, headers: headers) }, delay: delay)
+        self.mockEvery(matcher, delay: delay) { _ in return .Success(statusCode: statusCode, headers: headers, body: body) }
     }
 
     /**
-     The next request matching `expression` will successfully return the result of `body`, a method where the matches sections of the url as passed in as parameters.
+     The next request matching `expression` will successfully return the result of `response`, a method where the matched sections of the url are passed in as parameters.
 
      - parameter expression: The regular expression to compare incoming requests against
      - parameter HTTPMethod: The method to match against
      - parameter headers: The headers returned by the session data task
      - parameter statusCode: The status code (default=200) returned by the session data task
      - parameter delay: A artificial delay before the session data task starts to return response and data
-     - parameter body: Returns data the data to be  returned by the session data task. If this returns `nil` then the didRecieveData callback won't be called.
+     - parameter response: Returns data the data to be  returned by the session data task. If this returns `nil` then the didRecieveData callback won't be called.
      */
-    public class func mockSingle(expression: String, HTTPMethod: String = "GET", headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay, body: BodyFunction) throws {
+    public class func mockSingle(expression: String, HTTPMethod: String = "GET", delay: Double = DefaultDelay, response: MockResponseHandler) throws {
         let matcher = try SimpleRequestMatcher(expression: expression, method: HTTPMethod)
-        self.mockSingle(matcher, response: { (url: NSURL, extractions: [String]) in return MockResponse(body: body(extractions), statusCode: statusCode, headers: headers) }, delay: delay)
+        self.mockSingle(matcher, delay: delay, response: response)
     }
 
     /**
-     All subsequent requests matching `expression` will successfully return the result of `body`, a method where the matches sections of the url as passed in as parameters.
+     All subsequent requests matching `expression` will successfully return the result of `response`, a method where the matched sections of the url are passed in as parameters.
 
      - parameter expression: The regular expression to compare incoming requests against
      - parameter HTTPMethod: The method to match against
      - parameter headers: The headers returned by the session data task
      - parameter statusCode: The status code (default=200) returned by the session data task
      - parameter delay: A artificial delay before the session data task starts to return response and data
-     - parameter body: Returns data the data to be  returned by the session data task. If this returns `nil` then the didRecieveData callback won't be called.
+     - parameter response: Returns data the data to be  returned by the session data task. If this returns `nil` then the didRecieveData callback won't be called.
      */
-    public class func mockEvery(expression: String, HTTPMethod: String = "GET", headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay, body: BodyFunction) throws {
+    public class func mockEvery(expression: String, HTTPMethod: String = "GET", delay: Double = DefaultDelay, response: MockResponseHandler) throws {
         let matcher = try SimpleRequestMatcher(expression: expression, method: HTTPMethod)
-        self.mockEvery(matcher, response: { (url: NSURL, extractions: [String]) in return MockResponse(body: body(extractions), statusCode: statusCode, headers: headers) }, delay: delay)
+        self.mockEvery(matcher, delay: delay, response: response)
     }
-    
+
     /**
      Remove all mocks - NSURLSession will behave as if it had never been touched
      */
@@ -138,14 +132,14 @@ extension NSURLSession {
     //MARK: Private methods
     
     // Add a request matcher to the list of mocks
-    private class func mockSingle(matcher: RequestMatcher, response: MockResponseHandler, delay: Double) {
-        self.register.addMock(SingleSuccessSessionMock(matching: matcher, response: response, delay: delay))
+    private class func mockSingle(matcher: RequestMatcher, delay: Double, response: MockResponseHandler) {
+        self.register.addMock(SingleMockEntry(matching: matcher, response: response, delay: delay))
         swizzleIfNeeded()
     }
     
     // Add a request matcher to the list of mocks
-    private class func mockEvery(matcher: RequestMatcher, response: MockResponseHandler, delay: Double) {
-        self.register.addMock(SuccessSessionMock(matching: matcher, response: response, delay: delay))
+    private class func mockEvery(matcher: RequestMatcher, delay: Double, response: MockResponseHandler) {
+        self.register.addMock(MockEntry(matching: matcher, response: response, delay: delay))
         swizzleIfNeeded()
     }
 }